requirements.txt
pyTelegramBotAPI
google-generativeai
python-dotenv
psycopg2-binary
